requirements.txt
asgiref==3.7.2
Django==4.2.2
Pillow==9.5.0
python-decouple==3.8
sqlparse==0.4.4
tzdata==2023.3
